Create module Resourceable and clean up application controller
